Refined library. Enhanced the install script, so it can run from the …

+++ Install-MusicTools.ps1
@@ -10,6 +10,17 @@ and ensures that PowerShell 5.x is installed.
 Your Name
 #>
 
+#-------------------------------------------------------[Declarations]-------------------------------------------------------
+$installPS5Message = @"
+To install PowerShell 5.x:
+1. Download the Windows Management Framework 5.1 from the official Microsoft website:
+   https://www.microsoft.com/en-us/download/details.aspx?id=54616
+2. Follow the installation instructions provided on the download page.
+3. After installation, restart your system if prompted.
+
+Note: If you're using an older version of Windows (e.g., Windows 7/8), ensure your system meets the requirements listed on the download page.
+"@
+
 
 # Define paths
 $MusicToolsPath = "$env:USERPROFILE\MusicTools"
@@ -20,32 +31,52 @@ $IniFilePath = "$MusicToolsPath\UnrealCommanderToolbarButtons.ini"
 $TempIniFilePath = "$MusicToolsPath\UnrealCommanderToolbarButtons.temp.ini"
 $ConfigFilePath = "$MusicToolsPath\SourceDestinationPairs.txt"
 
-#-------------------------------------------------------[User Approval]-------------------------------------------------------
+#-------------------------------------------------------[PowerShell Checking]-------------------------------------------------
+# Ensure PowerShell 5.x is installed
+$PS5Path = "$env:SystemRoot\System32\WindowsPowerShell\v1.0\powershell.exe"
+if (Test-Path -Path $PS5Path) {
+    # Check the version of the executable
+    $PS5Version = & $PS5Path -Command { $PSVersionTable.PSVersion.Major }
 
-$UserInput = Read-Host "This script will download and configure MusicTools in '$MusicToolsPath' directory. Do you want to continue? Type 'Yes' to proceed"
+    if ($PS5Version -eq 5) {
+        Write-Host "`nPowerShell 5.x is installed and available at $PS5Path."
+    } else {
+        Write-Host -NoNewline "The PowerShell executable at '" 
+        Write-Host -NoNewline $PS5Path -ForegroundColor Red 
+        Write-Host "' is not version 5. Detected version: $PS5Version."
+        Write-Host $installPS5Message
+        exit 1
+    }
+} else {
+    Write-Host -NoNewline "`nPowerShell 5.x executable not found at '" 
+    Write-Host -NoNewline $PS5Path -ForegroundColor Red 
+    Write-Host "'. Please ensure it is installed."
+    Write-Host $installPS5Message
+    exit 1
+}
+
+
+#-------------------------------------------------------[User Approval]-------------------------------------------------------
+Write-Host "`nWelcome to the MusicTools installation script!" -ForegroundColor Green
+if(-not (Test-Path -Path $MusicToolsPath)) {
+    Write-Host -NoNewline "This script will create the '"
+    Write-Host -NoNewline $MusicToolsPath -ForegroundColor Red
+    Write-Host "' directory."
+} else {
+    Write-Host "The MusicTools folder '$MusicToolsPath' already exists." -ForegroundColor Green
+}
+
+Write-Host "The MusicTools repository will be downloaded and configured." -ForegroundColor Green
+Write-Host -NoNewline "Do you want to continue? Type exactly '"
+Write-Host -NoNewline -ForegroundColor Red "Yes"
+Write-Host -NoNewline "' to proceed: "
+$UserInput = Read-Host 
 
 if ($UserInput -ceq "Yes") {
-    Write-Host "Installation requested." -ForegroundColor Green
+    Write-Host "Installation requested. Directory '$MusicToolsPath' will be used." -ForegroundColor Green
 } else {
     Write-Host "Installation aborted by the user." -ForegroundColor Red
     exit 0
-}
-
-# Ensure PowerShell 5.x is installed
-if ($PSVersionTable.PSVersion.Major -eq 5) {
-    Write-Host "PowerShell 5.x detected."
-} else {
-    Write-Warning "PowerShell 5.x is required. Detected version: $($PSVersionTable.PSVersion)."
-    Write-Host @"
-To install PowerShell 5.x:
-1. Download the Windows Management Framework 5.1 from the official Microsoft website:
-   https://www.microsoft.com/en-us/download/details.aspx?id=54616
-2. Follow the installation instructions provided on the download page.
-3. After installation, restart your system if prompted.
-
-Note: If you're using an older version of Windows (e.g., Windows 7/8), ensure your system meets the requirements listed on the download page.
-"@
-    exit 1
 }
 
 # Step 1: Ensure the MusicTools folder exists
@@ -91,7 +122,7 @@ if (-not (Test-Path -Path $IniFilePath)) {
     Write-Host "INI file created and updated successfully."
 } else {
     Write-Host "INI file already exists. Preserving user-defined settings."
-    Remove-Item -Path $TempIniFilePath
+    if(Test-Path $TempIniFilePath){Remove-Item -Path $TempIniFilePath -Force}
 }
 
 # Step 5: Set up the config file using Set-ConfigFile
